requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
idna==3.10
lxml==5.3.0
multidict==6.1.0
outcome==1.3.0.post0
pyaes==1.6.1
pyasn1==0.6.1
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
Telethon==1.36.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.13.0
